#67 global filters on overview (#84)
https://github.com/ansible/insights-frontend/issues/67
https://trello.com/c/Hif3NWoq/167-all-systems-dropdown-is-also-not-working
