+++ app/js/components/groupSelect/groupSelect.directive.js
@@ -2,30 +2,14 @@
 
 var componentsModule = require('../');
 var isEmpty = require('lodash/isEmpty');
-const includes = require('lodash/includes');
-
-// group-select is disabled in these states
-const DISABLED_STATES = [
-    'app.config',
-    'app.digests',
-    'app.overview',
-    'app.rules'
-];
 
 /**
  * @ngInject
  */
-function groupSelectCtrl($scope, $rootScope, Group, Events, $state) {
+function groupSelectCtrl($scope, $rootScope, Group, Events) {
     Group.init();
     $scope.groups = Group.groups;
     $scope.group = Group.current();
-
-    function checkState () {
-        $scope.disabled = includes(DISABLED_STATES, $state.current.name);
-    }
-
-    $scope.$on('$stateChangeSuccess', checkState);
-    checkState();
 
     $scope.triggerChange = function (group) {
         $scope.group = group;
@@ -53,7 +37,8 @@ function groupSelect() {
         replace: true,
         controller: groupSelectCtrl,
         scope: {
-            round: '='
+            round: '=',
+            disabled: '<'
         }
     };
 }
